% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_read_fruits.R
\name{dissect_str}
\alias{dissect_str}
\title{function to read pdf with vegetable data
and helper function to dissect string into values
function pdf_read_fruts(fname) returns a data.frame if succesfull, NULL otherwise
function import utils and pdf_tools}
\usage{
dissect_str(string)
}
\description{
function to read pdf with vegetable data
and helper function to dissect string into values
function pdf_read_fruts(fname) returns a data.frame if succesfull, NULL otherwise
function import utils and pdf_tools
}
