% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_read.R
\name{pdf_read}
\alias{pdf_read}
\title{function to read and clean pdf files with meat prices
function reads the pdf as text
and reads the data into a data.frame
the function is simple enough and does not proceeds into many checks
the function returns a data.frame with 5 columns if it succeeds, NULL otherwise}
\usage{
pdf_read(link)
}
\description{
function to read and clean pdf files with meat prices
function reads the pdf as text
and reads the data into a data.frame
the function is simple enough and does not proceeds into many checks
the function returns a data.frame with 5 columns if it succeeds, NULL otherwise
}
