% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Read price data from website}
\usage{
get_data(
  start_date = "01-01-2010",
  end_date = "01-01-2024",
  cache_path = "./files/",
  use_cache = FALSE,
  food = "ALL"
)
}
\arguments{
\item{start_date}{start date,in \%d-\%m-\%Y, for data retrival}

\item{end_date}{end date,in  \%d-\%m-\%Y, for data retrival}

\item{cache_path}{path to store files}

\item{use_cache}{TRUE/FALSE whether to use a cache directory for storing files}

\item{food}{string ALL|MEAT|FRUITS|VEGETABLES to choose kind of food to report}
}
\value{
A dataframe with the requested data
}
\description{
This function reads the price data from the website, and returns
a data frame with the requested data.
}
